requirements-dev.txt
# Development dependencies — includes production deps + dev tooling
-r requirements.txt

# Code quality
ruff==0.15.5
bandit==1.9.4
pip-audit==2.10.0
mypy==1.19.1
django-stubs==5.2.9
django-migration-linter==6.0.0

# Testing
coverage==7.13.4
factory_boy==3.3.3
Faker==40.8.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-env
pytest-xdist==3.8.0

# Debug & profiling
django-debug-toolbar==6.2.0
django-silk==5.4.3
gprof2dot==2025.4.14

# Tunnelling
pyngrok==7.2.3

# Transitive (pinned for reproducibility)
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
